refactor: split credit notification config tab state
