+++ src/cook-web/src/app/admin/operations/credit-notifications/CreditNotificationConfigTab.tsx
@@ -1,11 +1,10 @@
-import { useState, type ReactNode } from 'react';
+import type { ReactNode } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useEnvStore } from '@/c-store';
 import type { EnvStoreState } from '@/c-types/store';
 import ErrorDisplay from '@/components/ErrorDisplay';
 import Loading from '@/components/loading';
 import { Button } from '@/components/ui/Button';
-import { toast } from '@/hooks/useToast';
 import { Input } from '@/components/ui/Input';
 import { Label } from '@/components/ui/Label';
 import {
@@ -21,7 +20,6 @@ import { resolveContactMode } from '@/lib/resolve-contact-mode';
 import type {
   AdminOperationCreditNotificationDryRunResponse,
   AdminOperationCreditNotificationPolicy,
-  AdminOperationCreditNotificationPolicyListItem,
   AdminOperationCreditNotificationPolicyResolvedLists,
   AdminOperationCreditNotificationTemplateOption,
   AdminOperationCreditNotificationTemplateSyncResponse,
@@ -30,7 +28,6 @@ import { CreditNotificationCreatorListsSection } from './CreditNotificationCreat
 import { CreditNotificationDryRunPanel } from './CreditNotificationDryRunPanel';
 import {
   CreditNotificationManagedListDialog,
-  type CreditNotificationManagedListType as ManagedListType,
 } from './CreditNotificationManagedListDialog';
 import {
   CreditNotificationTypeConfigCard,
@@ -42,20 +39,15 @@ import {
   CreditNotificationHelpTooltip as HelpTooltip,
 } from './CreditNotificationFormPrimitives';
 import {
-  formatListInput,
   isEstimatedDaysThreshold,
   isFixedThreshold,
   type KnownNotificationType,
   NOTIFICATION_TYPES,
-  normalizeIntegerInput,
-  normalizeListInputCharacters,
-  parseListInput,
-  readNumber,
 } from './creditNotificationUtils';
-
-type UpdatePolicy = (
-  updater: (draft: AdminOperationCreditNotificationPolicy) => void,
-) => void;
+import {
+  type UpdatePolicy,
+  useCreditNotificationConfigTabState,
+} from './useCreditNotificationConfigTabState';
 
 const TIMEZONE_OPTIONS = [
   'Asia/Shanghai',
@@ -66,148 +58,6 @@ const TIMEZONE_OPTIONS = [
   'Asia/Seoul',
   'UTC',
 ];
-const MAX_BLOCKED_CREATOR_IMPORT_COUNT = 50;
-const INVALID_SAMPLE_LIMIT = 5;
-const PHONE_MATCH_PATTERN = /(?:^|\D)(\d{11})(?!\d)/g;
-const EMAIL_MATCH_PATTERN = /[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}/g;
-const EMAIL_TEST_PATTERN = /^[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\.[A-Za-z]{2,}$/;
-const CREATOR_ID_PATTERN = /^[A-Za-z0-9_-]{4,80}$/;
-
-const isSmsMobileIdentifier = (value: string) => {
-  const normalized = value.trim().replace(/^\+/, '');
-  return /^\d{5,20}$/.test(normalized);
-};
-
-const findPhoneIdentifiers = (value: string): string[] => {
-  const matches: string[] = [];
-  const pattern = new RegExp(PHONE_MATCH_PATTERN.source, 'g');
-  let match = pattern.exec(value);
-  while (match) {
-    if (match[1]) {
-      matches.push(match[1]);
-    }
-    match = pattern.exec(value);
-  }
-  return matches;
-};
-
-const findEmailIdentifiers = (value: string): string[] =>
-  Array.from(value.matchAll(EMAIL_MATCH_PATTERN)).map(match =>
-    match[0].toLowerCase(),
-  );
-
-const splitIdentifierTokens = (value: string): string[] =>
-  value
-    .split(/[\s,;|]+/)
-    .map(item => item.trim())
-    .filter(Boolean);
-
-const trimInvalidDisplay = (value: string): string =>
-  value.replace(/^[\s,;|]+|[\s,;|]+$/g, '').trim();
-
-const parseBlockedCreatorInput = (
-  value: string,
-  contactMode: 'email' | 'phone',
-) => {
-  const normalized = normalizeListInputCharacters(value).replace(/\t/g, ' ');
-  const creatorBids: string[] = [];
-  const mobiles: string[] = [];
-  const invalidItems: string[] = [];
-  const addToken = (token: string) => {
-    if (isSmsMobileIdentifier(token)) {
-      mobiles.push(token);
-      return;
-    }
-    if (EMAIL_TEST_PATTERN.test(token)) {
-      creatorBids.push(token.toLowerCase());
-      return;
-    }
-    if (CREATOR_ID_PATTERN.test(token)) {
-      creatorBids.push(token);
-      return;
-    }
-    invalidItems.push(token);
-  };
-
-  normalized.split(/\r?\n/).forEach(rawLine => {
-    const line = trimInvalidDisplay(rawLine);
-    if (!line) {
-      return;
-    }
-    if (/[;,]/.test(line)) {
-      splitIdentifierTokens(line).forEach(addToken);
-      return;
-    }
-    const contactMatches =
-      contactMode === 'email'
-        ? findEmailIdentifiers(line)
-        : findPhoneIdentifiers(line);
-    if (contactMatches.length > 0) {
-      if (contactMode === 'email') {
-        creatorBids.push(...contactMatches);
-        return;
-      }
-      mobiles.push(...contactMatches);
-      return;
-    }
-
-    splitIdentifierTokens(line).forEach(addToken);
-  });
-
-  return {
-    creatorBids: Array.from(new Set(creatorBids)),
-    invalidItems: Array.from(new Set(invalidItems)),
-    mobiles: Array.from(new Set(mobiles)),
-  };
-};
-
-const mergeIdentifierLists = (...lists: string[][]) =>
-  Array.from(
-    new Set(
-      lists.flatMap(list => list.map(item => item.trim()).filter(Boolean)),
-    ),
-  );
-
-const buildListDetails = (
-  identifiers: string[],
-  resolvedItems: AdminOperationCreditNotificationPolicyListItem[] = [],
-) => {
-  const resolvedByIdentifier = new Map(
-    resolvedItems.map(item => [item.identifier, item]),
-  );
-  return identifiers.map(identifier => ({
-    identifier,
-    creator_bid: resolvedByIdentifier.get(identifier)?.creator_bid || '',
-    mobile: resolvedByIdentifier.get(identifier)?.mobile || '',
-    email: resolvedByIdentifier.get(identifier)?.email || '',
-    nickname: resolvedByIdentifier.get(identifier)?.nickname || '',
-  }));
-};
-
-const filterListDetails = (
-  items: AdminOperationCreditNotificationPolicyListItem[],
-  keyword: string,
-  contactMode: 'email' | 'phone',
-) => {
-  const normalized = keyword.trim().toLowerCase();
-  if (!normalized) {
-    return items;
-  }
-  return items.filter(item => {
-    const contactValue =
-      contactMode === 'email'
-        ? item.email || item.identifier
-        : item.mobile || item.identifier;
-    const isContactSearch =
-      contactMode === 'email'
-        ? normalized.includes('@')
-        : /^\d+$/.test(normalized);
-    if (isContactSearch) {
-      return contactValue.toLowerCase() === normalized;
-    }
-    return item.nickname.toLowerCase().includes(normalized);
-  });
-};
 
 function ConfigCard({
   title,
@@ -288,201 +138,45 @@ export function CreditNotificationConfigTab({
     loginMethodsEnabled,
     defaultLoginMethod,
   );
-  const [openTemplatePicker, setOpenTemplatePicker] = useState<
-    Partial<Record<KnownNotificationType, boolean>>
-  >({});
-  const [editingTemplateTypes, setEditingTemplateTypes] = useState<
-    Partial<Record<KnownNotificationType, boolean>>
-  >({});
-  const [templateInputValues, setTemplateInputValues] = useState<
-    Partial<Record<KnownNotificationType, string>>
-  >({});
-  const [listInputValues, setListInputValues] = useState<
-    Partial<Record<string, string>>
-  >({});
-  const [blockedCreatorInput, setBlockedCreatorInput] = useState('');
-  const [integerInputValues, setIntegerInputValues] = useState<
-    Partial<Record<string, string>>
-  >({});
-  const [managedListDialog, setManagedListDialog] =
-    useState<ManagedListType | null>(null);
-  const [managedListSearch, setManagedListSearch] = useState('');
+  const {
+    addBlockedCreators,
+    blockedCreatorIdentifiers,
+    blockedCreatorInput,
+    closeManagedListDialog,
+    editingTemplateTypes,
+    filteredManagedListDetails,
+    finishIntegerInput,
+    finishListInput,
+    getIntegerInputValue,
+    getListInputValue,
+    managedListCanDelete,
+    managedListDialog,
+    managedListSearch,
+    managedListTitle,
+    openManagedListDialog,
+    openTemplatePicker,
+    optedOutCreatorIdentifiers,
+    removeBlockedCreator,
+    setBlockedCreatorInput,
+    setEditingTemplateTypes,
+    setManagedListSearch,
+    setOpenTemplatePicker,
+    setTemplateInputValues,
+    templateInputValues,
+    updateIntegerInput,
+    updateListInput,
+  } = useCreditNotificationConfigTabState({
+    contactMode,
+    policy,
+    resolvedLists,
+    updatePolicy,
+    t,
+  });
   const lowBalanceThresholds = policy.types.low_balance.thresholds || [];
   const fixedLowBalanceThresholds =
     lowBalanceThresholds.filter(isFixedThreshold);
   const estimatedDaysThreshold =
     lowBalanceThresholds.find(isEstimatedDaysThreshold) || null;
-  const blockedCreatorIdentifiers = mergeIdentifierLists(
-    policy.blacklist.creator_bids,
-    policy.blacklist.mobiles,
-  );
-  const optedOutCreatorIdentifiers = mergeIdentifierLists(
-    policy.opt_out.creator_bids,
-    policy.opt_out.mobiles,
-  );
-  const blockedCreatorDetails = buildListDetails(
-    blockedCreatorIdentifiers,
-    resolvedLists.blacklist?.items || [],
-  );
-  const optedOutCreatorDetails = buildListDetails(
-    optedOutCreatorIdentifiers,
-    resolvedLists.opt_out?.items || [],
-  );
-
-  const getListInputValue = (key: string, value: string[]) =>
-    listInputValues[key] ?? formatListInput(value);
-
-  const updateListInput = (
-    key: string,
-    value: string,
-    commit: (normalized: string) => void,
-  ) => {
-    const inputValue = normalizeListInputCharacters(value);
-    setListInputValues(current => ({
-      ...current,
-      [key]: inputValue,
-    }));
-    commit(inputValue);
-  };
-
-  const finishListInput = (key: string, value: string) => {
-    setListInputValues(current => {
-      const next = { ...current };
-      next[key] = formatListInput(parseListInput(value));
-      return next;
-    });
-  };
-
-  const getIntegerInputValue = (key: string, value: number) =>
-    integerInputValues[key] ?? String(value);
-
-  const updateIntegerInput = (
-    key: string,
-    value: string,
-    fallback: number,
-    commit: (value: number) => void,
-  ) => {
-    const normalized = normalizeIntegerInput(value);
-    setIntegerInputValues(current => ({
-      ...current,
-      [key]: normalized,
-    }));
-    commit(readNumber(normalized, fallback));
-  };
-
-  const finishIntegerInput = (key: string, value: number) => {
-    setIntegerInputValues(current => {
-      const next = { ...current };
-      next[key] = String(value);
-      return next;
-    });
-  };
-
-  const openManagedListDialog = (type: ManagedListType) => {
-    setManagedListDialog(type);
-    setManagedListSearch('');
-  };
-
-  const removeBlockedCreator = (identifier: string) => {
-    updatePolicy(draft => {
-      draft.blacklist.creator_bids = draft.blacklist.creator_bids.filter(
-        item => item !== identifier,
-      );
-      draft.blacklist.mobiles = draft.blacklist.mobiles.filter(
-        item => item !== identifier,
-      );
-    });
-  };
-
-  const addBlockedCreators = () => {
-    const normalized = normalizeListInputCharacters(blockedCreatorInput);
-    const { creatorBids, invalidItems, mobiles } = parseBlockedCreatorInput(
-      normalized,
-      contactMode,
-    );
-    if (invalidItems.length > 0) {
-      const sample = invalidItems.slice(0, INVALID_SAMPLE_LIMIT).join(', ');
-      toast({
-        title: t(
-          'module.operationsCreditNotifications.config.listDialog.invalidBlockedCreators',
-          {
-            values:
-              invalidItems.length > INVALID_SAMPLE_LIMIT
-                ? `${sample}...`
-                : sample,
-          },
-        ),
-        variant: 'destructive',
-      });
-      return;
-    }
-    if (creatorBids.length === 0 && mobiles.length === 0) {
-      if (blockedCreatorInput) {
-        setBlockedCreatorInput('');
-      }
-      return;
-    }
-    const existingCreatorBids = new Set(policy.blacklist.creator_bids);
-    const existingMobiles = new Set(policy.blacklist.mobiles);
-    const nextCreatorBids = creatorBids.filter(
-      item => !existingCreatorBids.has(item),
-    );
-    const nextMobiles = mobiles.filter(item => !existingMobiles.has(item));
-    const addedCount = nextCreatorBids.length + nextMobiles.length;
-    if (addedCount > MAX_BLOCKED_CREATOR_IMPORT_COUNT) {
-      toast({
-        title: t(
-          'module.operationsCreditNotifications.config.listDialog.blockedCreatorLimit',
-          { count: MAX_BLOCKED_CREATOR_IMPORT_COUNT },
-        ),
-        variant: 'destructive',
-      });
-      return;
-    }
-    if (addedCount === 0) {
-      setBlockedCreatorInput('');
-      toast({
-        title: t(
-          'module.operationsCreditNotifications.config.listDialog.duplicateBlockedCreators',
-        ),
-      });
-      return;
-    }
-    updatePolicy(draft => {
-      draft.blacklist.creator_bids = Array.from(
-        new Set([...draft.blacklist.creator_bids, ...nextCreatorBids]),
-      );
-      draft.blacklist.mobiles = Array.from(
-        new Set([...draft.blacklist.mobiles, ...nextMobiles]),
-      );
-    });
-    setBlockedCreatorInput('');
-    toast({
-      title: t(
-        'module.operationsCreditNotifications.config.listDialog.addedBlockedCreators',
-        { count: addedCount },
-      ),
-    });
-  };
-
-  const managedListDetails =
-    managedListDialog === 'blocked'
-      ? blockedCreatorDetails
-      : optedOutCreatorDetails;
-  const filteredManagedListDetails = filterListDetails(
-    managedListDetails,
-    managedListSearch,
-    contactMode,
-  );
-  const managedListTitle =
-    managedListDialog === 'blocked'
-      ? t(
-          'module.operationsCreditNotifications.config.fields.blockedCreatorList',
-        )
-      : t(
-          'module.operationsCreditNotifications.config.fields.optedOutCreators',
-        );
-  const managedListCanDelete = managedListDialog === 'blocked';
 
   if (configLoading && !configLoaded) {
     return (
@@ -1019,10 +713,7 @@ export function CreditNotificationConfigTab({
           search={managedListSearch}
           onSearchChange={setManagedListSearch}
           onRemove={removeBlockedCreator}
-          onClose={() => {
-            setManagedListDialog(null);
-            setManagedListSearch('');
-          }}
+          onClose={closeManagedListDialog}
         />
       </div>
